+++ src/electrifier/Views/ShellPage.xaml
@@ -156,7 +156,7 @@
                                   Glyph="&#xeb3c;" />
                     </NavigationViewItem.Icon>
                 </NavigationViewItem>
-                <NavigationViewItem x:Uid="Shell_People"
+                <NavigationViewItem x:Uid="ShellPage_People"
                                     IsEnabled="False">
                     <NavigationViewItem.Icon>
                         <FontIcon FontFamily="{StaticResource SymbolThemeFontFamily}"
